Range workaround

+++ segme/common/impfunc.py
@@ -14,8 +14,9 @@ def make_coords(batch, height, width, dtype=None, name=None):
         height_ = 1.0 / ops.cast(height, dtype)
         width_ = 1.0 / ops.cast(width, dtype)
 
-        vertical = height_ - 1.0 + 2 * height_ * ops.arange(height, dtype=dtype)
-        horizontal = width_ - 1.0 + 2 * width_ * ops.arange(width, dtype=dtype)
+        # TODO https://github.com/tensorflow/tensorflow/issues/80039
+        vertical = height_ - 1.0 + 2 * height_ * ops.cast(ops.arange(height, dtype="float32"), dtype)
+        horizontal = width_ - 1.0 + 2 * width_ * ops.cast(ops.arange(width, dtype="float32"), dtype)
 
         mesh = ops.meshgrid(vertical, horizontal, indexing="ij")
         join = ops.stack(mesh, axis=-1)
